+++ src/transaction/Transaction.ts
@@ -12,7 +12,8 @@ import Spend from '../script/Spend.js'
 import ChainTracker from './ChainTracker.js'
 import { defaultBroadcaster } from './broadcasters/DefaultBroadcaster.js'
 import { defaultChainTracker } from './chaintrackers/DefaultChainTracker.js'
-import { BEEF_MAGIC } from './Beef.js'
+import { ATOMIC_BEEF, BEEF_MAGIC } from './Beef.js'
+import P2PKH from '../script/templates/P2PKH.js'
 
 /**
  * Represents a complete Bitcoin transaction. This class encapsulates all the details
@@ -119,8 +120,8 @@ export default class Transaction {
     const reader = new Reader(beef)
     // Read the Atomic BEEF prefix
     const prefix = reader.readUInt32LE()
-    if (prefix !== 0x01010101) {
-      throw new Error(`Invalid Atomic BEEF prefix. Expected 0x01010101, received ${prefix.toString(16)}.`)
+    if (prefix !== ATOMIC_BEEF) {
+      throw new Error(`Invalid Atomic BEEF prefix. Expected 0x01010101, received 0x${prefix.toString(16)}.`)
     }
 
     // Read the subject TXID
@@ -455,7 +456,28 @@ export default class Transaction {
    */
   addOutput (output: TransactionOutput): void {
     this.cachedHash = undefined
+    if (!output.change) {
+      if (typeof output.satoshis === 'undefined') throw new Error('either satoshis must be defined or change must be set to true')
+      if (output.satoshis <= 0) throw new Error('satoshis must be a positive integer or zero')
+    }
+    if (!output.lockingScript) throw new Error('lockingScript must be defined')
     this.outputs.push(output)
+  }
+
+  /**
+   * Adds a new P2PKH output to the transaction.
+   *
+   * @param {number[] | string} address - The P2PKH address of the output.
+   * @param {number} [satoshis] - The number of satoshis to send to the address - if not provided, the output is considered a change output.
+   *
+   */
+  addP2PKHOutput (address: number[] | string, satoshis?: number): void {
+    const lockingScript = new P2PKH().lock(address)
+    if (typeof satoshis === 'undefined') { return this.addOutput({ lockingScript, change: true }) }
+    this.addOutput({
+      lockingScript,
+      satoshis
+    })
   }
 
   /**